[Bronze II] Title: Hashing, Time: 68 ms, Memory: 11540 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -1,6 +1,6 @@
 import java.io.*;
 class Main{
-    static int ft(int n,int dt){
+    static long ft(long n,int dt){
         if(dt==0) return n;
         n=n*31%1234567891;
         return ft(n,dt-1);
@@ -9,7 +9,7 @@ class Main{
         BufferedReader I=new BufferedReader(new InputStreamReader(System.in));
         int n=Integer.parseInt(I.readLine());
         String s=I.readLine();
-        int res=0;
+        long res=0;
         for(int i=0;i<n;i++){
             res+=ft(s.charAt(i)-96,i)%1234567891;
         }
